Update rollup data writer
