+++ src/test/scala/io/qbeast/spark/delta/DeltaRollupDataWriterTest.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.qbeast.spark.delta.writer
+package io.qbeast.spark.delta
 
 import io.qbeast.core.model.BroadcastedTableChanges
 import io.qbeast.core.model.IndexStatus
