+++ src/test/scala/io/qbeast/spark/writer/BlockWriterTest.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.qbeast.spark.delta.writer
+package io.qbeast.spark.writer
 
 import io.qbeast.core.model.CubeId
 import io.qbeast.spark.delta.QbeastFileUtils
